+++ CHANGELOG.md
@@ -4,9 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.5] - 2024-02-26
+### Fixed
+- Also report async errors and custom error states
+- Fix build number not being set correctly
+
 ## [1.0.4] - 2024-02-24
 ### Fixed
-- Fix provider errors not beeing sent to sentry
+- Fix provider errors not being sent to sentry
 
 ## [1.0.3] - 2024-02-23
 ### Added
@@ -28,6 +33,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Initial release
 
+[1.0.5]: https://github.com/Skycoder42/wear_tasks/compare/v1.0.4...v1.0.5
 [1.0.4]: https://github.com/Skycoder42/wear_tasks/compare/v1.0.3...v1.0.4
 [1.0.3]: https://github.com/Skycoder42/wear_tasks/compare/v1.0.2...v1.0.3
 [1.0.2]: https://github.com/Skycoder42/wear_tasks/compare/v1.0.1...v1.0.2